finished it only give back an error because the database is empty
