+++ src/routes/adminRoutes.js
@@ -1,9 +1,8 @@
-import AdminService from '../sevices/adminService';
-const express = require('express');
+const AdminController = require("../controllers/adminController"); // Use require instead of import
+const express = require("express");
+
 const router = express.Router();
 
-const adminController = new AdminController
+router.post("/role", (req, res) => AdminController.changeUserRole(req, res));
 
-router.post('/role', (req, res) => adminController.changeUserRole(req ,res));
-
-module.exports = router;
+module.exports = router; // Use module.exports instead of export default
